+++ server/src/routes/admin/registrations/createRegistration.js
@@ -10,7 +10,7 @@ module.exports = async (req, res, next) => {
       return next(Boom.badRequest("This user does not exist."));
     }
 
-    let camp = await Camp.findOne({ _id: req.body.camp });
+    let camp = await Camp.findOne({ _id: req.body.camp, archived: false });
     if (!camp) {
       return next(Boom.badRequest("This camp does not exist."));
     }
@@ -30,8 +30,10 @@ module.exports = async (req, res, next) => {
       deposit: req.body.deposit || false,
       paid: req.body.paid || false,
       waitlist: req.body.waitlist || false,
+      spaceSaved: req.body.spaceSaved || false,
       created: Date.now(),
-      user: req.body.user
+      user: req.body.user,
+      archived: false
     });
 
     await registration.save();